% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titanic_gender_model.R
\docType{data}
\name{titanic_gender_model}
\alias{titanic_gender_model}
\title{Titanic gender model data.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
http://www.something.com
}
\usage{
titanic_gender_model
}
\description{
Description.
}
\examples{
titanic_gender_model
}
\keyword{datasets}

