% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titanic_test.R
\docType{data}
\name{titanic_test}
\alias{titanic_test}
\title{Titanic test data.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
http://www.something.com
}
\usage{
titanic_test
}
\description{
Description.
}
\examples{
titanic_test
}
\keyword{datasets}

