% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titanic_train.R
\docType{data}
\name{titanic_train}
\alias{titanic_train}
\title{Titanic train data.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
http://www.something.com
}
\usage{
titanic_train
}
\description{
Description.
}
\examples{
titanic_train
}
\keyword{datasets}

