% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titanic_gender_model.R
\docType{data}
\name{titanic_gender_model}
\alias{titanic_gender_model}
\title{Titanic gender model data.}
\format{Data frame with columns
\describe{
\item{PassengerId}{Passenger ID}
\item{Survived}{Passenger Survival Indicator}
}}
\source{
https://www.kaggle.com/c/titanic/data
}
\usage{
titanic_gender_model
}
\description{
Titanic gender model data.
}
\examples{
titanic_gender_model
}
\keyword{datasets}

