% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_test.R
\docType{data}
\name{titanic_test}
\alias{titanic_test}
\title{Titanic test data.}
\format{Data frame with columns
\describe{
\item{PassengerId}{Passenger ID}
\item{Pclass}{Passenger Class}
\item{Name}{Name}
\item{Sex}{Sex}
\item{Age}{Age}
\item{SibSp}{Number of Siblings/Spouses Aboard}
\item{Parch}{Number of Parents/Children Aboard}
\item{Ticket}{Ticket Number}
\item{Fare}{Passenger Fare}
\item{Cabin}{Cabin}
\item{Embarked}{Port of Embarkation}
}}
\source{
https://www.kaggle.com/c/titanic/data
}
\usage{
titanic_test
}
\description{
Titanic test data.
}
\examples{
  titanic_test
}
\keyword{datasets}
