% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titanic_train.R
\docType{data}
\name{titanic_train}
\alias{titanic_train}
\title{Titanic train data.}
\format{Data frame with columns
\describe{
\item{PassengerId}{Passenger ID}
\item{Survived}{Passenger Survival Indicator}
\item{Pclass}{Passenger Class}
\item{Name}{Name}
\item{Sex}{Sex}
\item{Age}{Age}
\item{SibSp}{Number of Siblings/Spouses Aboard}
\item{Parch}{Number of Parents/Children Aboard}
\item{Ticket}{Ticket Number}
\item{Fare}{Passenger Fare}
\item{Cabin}{Cabin}
\item{Embarked}{Port of Embarkation}
}}
\source{
https://www.kaggle.com/c/titanic/data
}
\usage{
titanic_train
}
\description{
Titanic train data.
}
\examples{
titanic_train
}
\keyword{datasets}

